backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
sqlite3
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.14.0
requests==2.31.0
python-dotenv==1.0.0
plotly==5.17.0
folium==0.15.0
geopy==2.4.0
rasterio==1.3.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pytest==7.4.3